requirements.txt
torch>=2.2.0
transformers>=5.2.0,<5.3.0
datasets>=4.6.1
accelerate>=1.11.0
peft>=0.18.1
trl>=0.27.2
bitsandbytes>=0.49.2
huggingface_hub>=0.36.2
pyyaml>=6.0.2
sentencepiece>=0.2.0
tqdm>=4.67.0
wandb>=0.25.0
